refactor(shared): Extract ApiEnvelope to break import cycle
