+++ src/shared/http/index.ts
@@ -13,7 +13,8 @@
  * @see {@link module:shared/http/create-api-response-util}
  */
 
-export * from './create-api-response-util'
-export * from './api-response-serialize-util'
-export * from './with-error-handling'
-export * from './with-validation'
+export type { ApiEnvelope } from './api-envelope'
+export { createSuccessResponse, createErrorResponse } from './create-api-response-util'
+export { jsonResponse, mergeResponseHeaders } from './api-response-serialize-util'
+export { withErrorHandling } from './with-error-handling'
+export { withZodValidation } from './with-validation'
